pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veriflow-agent"
version = "0.1.0"
description = "Agent-based RTL design pipeline using LangGraph"
readme = "readme_first.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "VeriFlow Team", email = "team@veriflow.ai"}
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "anthropic>=0.30.0",
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "watchdog>=3.0",
    "gradio>=4.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "python-telegram-bot>=21.0",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.3",
    "black>=24.0",
]

[project.scripts]
veriflow-agent = "veriflow_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
veriflow_agent = ["prompts/*.md", "gateway/static/*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
